♻️ refactor(test): consolidate test-utils and testkit into unified testing directory
Consolidate test-utils/ and testkit/ directories into a unified testing/ structure
following DDD principles with proper layer separation:

• testing/infra/ - Infrastructure layer test utilities (database, process, runtime)
• testing/domain/ - Domain layer test utilities (fakes, stubs, integration helpers)
• testing/ - Convenience barrel exports for both layers

Key changes:
- Move test-utils/* → testing/infra/*
- Move testkit domain utilities → testing/domain/*
- Update all import paths to use @api/testing/* aliases
- Fix migration paths in drizzle.config.test.ts
- Update tsconfig exclude patterns
- Remove redundant testkit/ directory

All tests pass with improved test utility organization and maintainability.
