+++ apps/api/drizzle.config.test.ts
@@ -1,8 +1,8 @@
 import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
-  schema: './test-utils/db/schema.ts',
-  out: './test-utils/db/migrations',
+  schema: './testing/infra/db/schema.ts',
+  out: './testing/infra/db/migrations',
   dialect: 'postgresql',
   // No dbCredentials needed - we only generate SQL, not apply it
   verbose: true,
