+++ apps/api/src/system/health/health.integration.test.ts
@@ -1,5 +1,5 @@
 import { app } from '@api/index';
-import { setupTestDatabase } from '@api/test-utils/integration-helpers';
+import { setupTestDatabase } from '@api/testing/domain';
 import { describe, expect, it } from 'vitest';
 
 describe('Health check endpoints', () => {
