+++ apps/api/testing/infra/errors/database-error.ts
