+++ apps/api/tests/fakes/index.ts
@@ -3,6 +3,6 @@
  * @fileoverview Re-exports all fake implementations
  */
 
-export { FakeQuizRepository } from '../../test-utils/fakes/FakeQuizRepository';
-export { FakeUnitOfWork } from '../../test-utils/fakes/FakeUnitOfWork';
-export { FakeUserRepository } from '../../test-utils/fakes/FakeUserRepository';
+export { FakeQuizRepository } from '@api/testing/domain/fakes/FakeQuizRepository';
+export { FakeUnitOfWork } from '@api/testing/domain/fakes/FakeUnitOfWork';
+export { FakeUserRepository } from '@api/testing/domain/fakes/FakeUserRepository';
