+++ apps/api/tests/vitest.config.integration.ts
@@ -25,7 +25,7 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { defineConfig, loadEnv } from 'vite';
 import tsconfigPaths from 'vite-tsconfig-paths';
-import { mapTestEnvironmentVariables } from '../test-utils/vitest';
+import { mapTestEnvironmentVariables } from '../testing/infra/vitest';
 
 // Get root directory path
 const __dirname = dirname(fileURLToPath(import.meta.url));
